allow exceptions in DeleteNonReferencedReferencesUpdater

+++ cdmlib-io/src/main/java/eu/etaxonomy/cdm/io/operation/DeleteNonReferencedReferencesUpdater.java
@@ -8,6 +8,7 @@
 */
 package eu.etaxonomy.cdm.io.operation;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.stereotype.Component;
@@ -17,7 +18,9 @@ import eu.etaxonomy.cdm.io.common.CdmImportBase;
 import eu.etaxonomy.cdm.io.common.DefaultImportState;
 import eu.etaxonomy.cdm.io.operation.config.DeleteNonReferencedReferencesConfigurator;
 import eu.etaxonomy.cdm.model.agent.TeamOrPersonBase;
+import eu.etaxonomy.cdm.model.common.IdentifiableSource;
 import eu.etaxonomy.cdm.model.reference.Reference;
+import eu.etaxonomy.cdm.persistence.query.OrderHint;
 
 /**
  * @author k.luther
@@ -31,8 +34,11 @@ public class DeleteNonReferencedReferencesUpdater extends CdmImportBase<DeleteNo
     @Override
 	protected void doInvoke(DefaultImportState<DeleteNonReferencedReferencesConfigurator> state) {
 
+        List<OrderHint> orderHint = new ArrayList<>();
+        orderHint.add(OrderHint.ORDER_BY_ID);
+
 		if (state.getConfig().isDoAuthors()){
-			List<TeamOrPersonBase> authors =getAgentService().list(TeamOrPersonBase.class, null, null, null, null);
+			List<TeamOrPersonBase> authors =getAgentService().list(TeamOrPersonBase.class, null, null, orderHint, null);
 
 			int deleted = 0;
 			System.out.println("There are " + authors.size() + " authors");
@@ -44,32 +50,73 @@ public class DeleteNonReferencedReferencesUpdater extends CdmImportBase<DeleteNo
 					if (!result.isOk()){
 						System.out.println("Author " + author.getTitleCache() + " with id " + author.getId() + " could not be deleted.");
 						result = null;
+					}else{
+					    System.out.println("Deleted: " + author.getTitleCache() + "; id = " + author.getId());
 					}
 				}
 			}
 			System.out.println(deleted + " authors are deleted.");
 		}
+
+		List<String> propertyPath = new ArrayList<>();
+		propertyPath.add("sources.citation");
+		propertyPath.add("createdBy");
 		if (state.getConfig().isDoReferences()){
-			List<Reference> references =getReferenceService().list(Reference.class, null, null, null, null);
+			List<Reference> references =getReferenceService().list(Reference.class, null, null, orderHint, propertyPath);
 
 			int deleted = 0;
 			System.out.println("There are " + references.size() + " references");
 			for (Reference ref: references){
 				long refObjects = getCommonService().getReferencingObjectsCount(ref);
 				if (refObjects == 0) {
-				    DeleteResult result = getReferenceService().delete(ref);
-					deleted++;
-					if (!result.isOk()){
-						System.out.println("Reference " + ref.getTitle() + " with id " + ref.getId() + " could not be deleted.");
-						result = null;
-					}
+				    if (isIgnore(state, ref)){
+				        System.out.println("Ignore: " + ref.getId() + "\t" + ref.getType() + "\t" +ref.getTitleCache() + "\t" + ref.getCreated()+ "\t" +
+				                (ref.getCreatedBy() == null? "" : ref.getCreatedBy().getUsername()) + "\t" +
+				                ref.getUpdated() + "\t" +  getSources(ref));
+				    }else{
+				        DeleteResult result = getReferenceService().delete(ref);
+				        deleted++;
+				        if (!result.isOk()){
+				            System.out.println("Reference " + ref.getTitle() + " with id " + ref.getId() + " could not be deleted.");
+				            result = null;
+				        }else{
+//				            System.out.println("Deleted: " + ref.getTitleCache() + "; id = " + ref.getId());
+				        }
+				    }
 				}
 			}
 			System.out.println(deleted + " references are deleted.");
 		}
 	}
 
-	@Override
+    private boolean isIgnore(DefaultImportState<DeleteNonReferencedReferencesConfigurator> state, Reference ref) {
+        if (state.getConfig().isKeepReferencesWithTitle() && isNotBlank(ref.getTitle())
+                || state.getConfig().isKeepRisSources() && hasRISSource(ref)){
+            return true;
+        }else{
+            return false;
+        }
+    }
+
+    private String getSources(Reference ref) {
+        String result = "";
+        for (IdentifiableSource source : ref.getSources()){
+            result += source.getType() + ": " + (source.getCitation() == null? "" : source.getCitation().getTitleCache()) + "\t";
+        }
+        return result;
+    }
+
+    private boolean hasRISSource(Reference ref) {
+        for (IdentifiableSource source : ref.getSources()){
+            Reference citation = source.getCitation();
+            if (citation != null && citation.getTitleCache().startsWith("RIS Reference")){
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
 	protected boolean doCheck(DefaultImportState<DeleteNonReferencedReferencesConfigurator> state) {
 		return true;
 	}
